Refactor QRC20 * Replace Swap specific functions to the separate swap_ops.rs module
